requirements.txt
diffusers
pillow
PyYAML
tqdm
transformers
accelerate
